[emacs] Replace tide with lsp-mode
Emacs 27 supposedly comes with a boatload of improvements related to
json parsing (order of 10x improved), which is heavily used by lsp-mode.
In Emacs 26 lsp-mode felt sluggish on trivial projects, but now it seems
to handle the trivial projects.

The original tide code was not removed, since it hasn't been tested on a
big project (which frankly tide also struggles with). At least lsp-mode
feels a lot faster compared to tide on smaller projects atm.